Add a new option testing_mode, which will help the statistics test
Essentially in testing_mode we will output the string '[template]'
for statistics for things like date, memory usage etc. We don't
want the literal values to be compared during phpunit testing as
those values change from run to run. Where we don't want '[template]'
to be output is for instance the number of tags processed which
should be stable from run to run.